source/requirements.txt
Django==1.11
django-isbn-field==0.5.2
